New Timer for Player 0 implemented
It looks ahead to determine to distance to the next .ON cell and choses the maximum available length in the soundBuffer to schedule.

New Property in SounBuffers struct:  lengthOfBufferInWholeCells - if a sound file is 8000 samples long, and a 16th note duration is 5512,5 samples (e.g. @ 120 bpm), the buffer will be 2 16th notes long = 11.025 samples, so the whole file fits in. lengthOfBufferInWholeCells for this file will be 2.

SoundBuffers .normal are now a array(!) of AVAudioPCMBuffer with different lengths (1/16 of a bar to full bar) to chose from.

TODO: Change Timers for player 1,2,3 accordingly.

New MainVC property: cellsToWaitBeforeRescheduling0 = 0 (default). maximum would be 15